# Optimizations for read_po