… tests/messages/test_normalized_string.py