Inspired by #1199.

This PR improves parsing performance for a large PO file (https://projects.blender.org/blender/blender-manual-translations/raw/branch/main/fi/LC_MESSAGES/blender_manual.po) by approximately 1.4x on my machine.

On a corpus of 11,600 .po files (namely all that I had hanging around on my machine, heh) the speed-up is 1.30x:

```
$ hyperfine 'git checkout read-po-opt-nr && python readpobench_many.py' 'git checkout master && python readpobench_many.py'
Benchmark 1: git checkout read-po-opt-nr && python readpobench_many.py
  Time (mean ± σ):      8.176 s ±  0.075 s    [User: 7.683 s, System: 0.478 s]
  Range (min … max):    8.052 s …  8.311 s    10 runs

Benchmark 2: git checkout master && python readpobench_many.py
  Time (mean ± σ):     10.638 s ±  0.122 s    [User: 10.153 s, System: 0.469 s]
  Range (min … max):   10.491 s … 10.880 s    10 runs

Summary
  git checkout read-po-opt-nr && python readpobench_many.py ran
    1.30 ± 0.02 times faster than git checkout master && python readpobench_many.py
```

The only "breaking" change here is the one implemented by f988d79074b99bdb83f7096640292c92590f0384 – you can no longer pass a mixed iterable of byteses or strs into `read_po`. Typing-wise, you never could (`AnyStr` is either `bytes` or `str`, but not an union of them), but now it'll actually break.

Internally, `_NormalizedString` lost some of its lustre, but it's fine, since it is an internal helper no one should ever see.